+++ corbind-material/src/main/kotlin/ru/ldralighieri/corbind/material/SearchViewTransitionStateChanges.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Vladimir Raupov
+ * Copyright 2022 Vladimir Raupov
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,8 @@
 
 package ru.ldralighieri.corbind.material
 
-import android.view.View
 import androidx.annotation.CheckResult
-import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.android.material.search.SearchView
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
@@ -32,50 +31,49 @@ import kotlinx.coroutines.isActive
 import ru.ldralighieri.corbind.internal.corbindReceiveChannel
 
 /**
- * Perform an action on the slide offset events from [View] on [BottomSheetBehavior].
+ * Perform an action on the transition state change events on [SearchView].
  *
  * @param scope Root coroutine scope
  * @param capacity Capacity of the channel's buffer (no buffer by default)
  * @param action An action to perform
  */
-fun View.slides(
+fun SearchView.transitionStateChanges(
     scope: CoroutineScope,
     capacity: Int = Channel.RENDEZVOUS,
-    action: suspend (Float) -> Unit
+    action: suspend (SearchView.TransitionState) -> Unit
 ) {
-    val events = scope.actor<Float>(Dispatchers.Main.immediate, capacity) {
-        for (offset in channel) action(offset)
+    val events = scope.actor<SearchView.TransitionState>(Dispatchers.Main.immediate, capacity) {
+        for (state in channel) action(state)
     }
 
-    val behavior = getBottomSheetBehavior()
-    val callback = callback(scope, events::trySend)
-    behavior.addBottomSheetCallback(callback)
-    events.invokeOnClose { behavior.removeBottomSheetCallback(callback) }
+    val listener = listener(scope, events::trySend)
+    addTransitionListener(listener)
+    events.invokeOnClose { removeTransitionListener(listener) }
 }
 
 /**
- * Perform an action on the slide offset events from [View] on [BottomSheetBehavior], inside new
+ * Perform an action on the transition state change events on [SearchView], inside new
  * [CoroutineScope].
  *
  * @param capacity Capacity of the channel's buffer (no buffer by default)
  * @param action An action to perform
  */
-suspend fun View.slides(
+suspend fun SearchView.transitionStateChanges(
     capacity: Int = Channel.RENDEZVOUS,
-    action: suspend (Float) -> Unit
+    action: suspend (SearchView.TransitionState) -> Unit
 ) = coroutineScope {
-    slides(this, capacity, action)
+    transitionStateChanges(this, capacity, action)
 }
 
 /**
- * Create a channel which emits the slide offset events from [View] on [BottomSheetBehavior].
+ * Create a channel which emits the transition state change events on [SearchView].
  *
  * Example:
  *
  * ```
  * launch {
- *      bottomSheetBehavior.slides(scope)
- *          .consumeEach { /* handle slide offset */ }
+ *      searchView.transitionStateChanges(scope)
+ *          .consumeEach { /* handle transition state change even */ }
  * }
  * ```
  *
@@ -83,45 +81,38 @@ suspend fun View.slides(
  * @param capacity Capacity of the channel's buffer (no buffer by default)
  */
 @CheckResult
-fun View.slides(
+fun SearchView.transitionStateChanges(
     scope: CoroutineScope,
     capacity: Int = Channel.RENDEZVOUS
-): ReceiveChannel<Float> = corbindReceiveChannel(capacity) {
-    val behavior = getBottomSheetBehavior()
-    val callback = callback(scope, ::trySend)
-    behavior.addBottomSheetCallback(callback)
-    invokeOnClose { behavior.removeBottomSheetCallback(callback) }
+): ReceiveChannel<SearchView.TransitionState> = corbindReceiveChannel(capacity) {
+    val listener = listener(scope, ::trySend)
+    addTransitionListener(listener)
+    invokeOnClose { removeTransitionListener(listener) }
 }
 
 /**
- * Create a flow which emits the slide offset events from [View] on [BottomSheetBehavior].
+ * Create a flow which emits the transition state change events on [SearchView].
  *
  * Example:
  *
  * ```
- * bottomSheetBehavior.slides()
- *      .onEach { /* handle slide offset */ }
+ * searchView.transitionStateChanges()
+ *      .onEach { /* handle transition state change event */ }
  *      .flowWithLifecycle(lifecycle)
  *      .launchIn(lifecycleScope) // lifecycle-runtime-ktx
  * ```
  */
 @CheckResult
-fun View.slides(): Flow<Float> = channelFlow {
-    val behavior = getBottomSheetBehavior()
-    val callback = callback(this, ::trySend)
-    behavior.addBottomSheetCallback(callback)
-    awaitClose { behavior.removeBottomSheetCallback(callback) }
+fun SearchView.transitionStateChanges(): Flow<SearchView.TransitionState> = channelFlow {
+    val listener = listener(this, ::trySend)
+    addTransitionListener(listener)
+    awaitClose { removeTransitionListener(listener) }
 }
 
 @CheckResult
-private fun callback(
+private fun listener(
     scope: CoroutineScope,
-    emitter: (Float) -> Unit
-) = object : BottomSheetBehavior.BottomSheetCallback() {
-
-    override fun onSlide(bottomSheet: View, slideOffset: Float) {
-        if (scope.isActive) { emitter(slideOffset) }
-    }
-
-    override fun onStateChanged(bottomSheet: View, newState: Int) = Unit
+    emitter: (SearchView.TransitionState) -> Unit
+) = SearchView.TransitionListener { _, _, newState ->
+    if (scope.isActive) { emitter(newState) }
 }